Rename automation.py to trigger.py
