+++ esphome/components/lvgl/trigger.py
